Adding Mapper, Reducer and SortKeys file.
Mapper takes initial means from mean file and reads mapper data. It then assigns all the data to one of the nearest clusters and ejects a key (Cluster) and Value(the data itself). Sortkeys.py is written to sort intermediate keys, so that we can test the functionality without actually using a mapReduce framework. The sorted keys are then sent to Reducer, and it
takes all the keys with similar cluster and emits a new mean.